+++ src/Entities/QueryString.php
@@ -34,8 +34,8 @@ class QueryString implements EntityInterface
     public static function load(array $data): static
     {
         return new static(
-            name: $data['name'] ?? throw InvalidArgumentException::missing('log.entries[].request[].queryString[].name'),
-            value: $data['value'] ?? throw InvalidArgumentException::missing('log.entries[].request[].queryString[].value'),
+            name: $data['name'] ?? throw InvalidArgumentException::missing('name', static::class),
+            value: $data['value'] ?? throw InvalidArgumentException::missing('value', static::class),
             comment: $data['comment'] ?? null,
         );
     }